Float -> Double

+++ src/main/java/com/bulletphysics/BulletGlobals.java
@@ -7,11 +7,11 @@
  * This software is provided 'as-is', without any express or implied warranty.
  * In no event will the authors be held liable for any damages arising from
  * the use of this software.
- * 
- * Permission is granted to anyone to use this software for any purpose, 
+ *
+ * Permission is granted to anyone to use this software for any purpose,
  * including commercial applications, and to alter it and redistribute it
  * freely, subject to the following restrictions:
- * 
+ *
  * 1. The origin of this software must not be misrepresented; you must not
  *    claim that you wrote the original software. If you use this software
  *    in a product, an acknowledgment in the product documentation would be
@@ -29,23 +29,23 @@ import cz.advel.stack.Stack;
 
 /**
  * Bullet global settings and constants.
- * 
+ *
  * @author jezek2
  */
 public class BulletGlobals {
-	
+
 	public static final boolean DEBUG = false;
-	
-	public static final float CONVEX_DISTANCE_MARGIN = 0.04f;
-	public static final float FLT_EPSILON = 1.19209290e-07f;
-	public static final float SIMD_EPSILON = FLT_EPSILON;
-	
-	public static final float SIMD_2_PI = 6.283185307179586232f;
-	public static final float SIMD_PI = SIMD_2_PI * 0.5f;
-	public static final float SIMD_HALF_PI = SIMD_2_PI * 0.25f;
-	public static final float SIMD_RADS_PER_DEG = SIMD_2_PI / 360f;
-	public static final float SIMD_DEGS_PER_RAD = 360f / SIMD_2_PI;
-	public static final float SIMD_INFINITY = Float.MAX_VALUE;
+
+	public static final double CONVEX_DISTANCE_MARGIN = 0.04f;
+	public static final double FLT_EPSILON = 1.19209290e-07f;
+	public static final double SIMD_EPSILON = FLT_EPSILON;
+
+	public static final double SIMD_2_PI = 6.283185307179586232f;
+	public static final double SIMD_PI = SIMD_2_PI * 0.5f;
+	public static final double SIMD_HALF_PI = SIMD_2_PI * 0.25f;
+	public static final double SIMD_RADS_PER_DEG = SIMD_2_PI / 360f;
+	public static final double SIMD_DEGS_PER_RAD = 360f / SIMD_2_PI;
+	public static final double SIMD_INFINITY = Double.MAX_VALUE;
 
 	////////////////////////////////////////////////////////////////////////////
 
@@ -60,11 +60,11 @@ public class BulletGlobals {
 	private ContactAddedCallback gContactAddedCallback;
 	private ContactProcessedCallback gContactProcessedCallback;
 
-	private float contactBreakingThreshold = 0.02f;
+	private double contactBreakingThreshold = 0.02f;
 	// RigidBody
-	private float deactivationTime = 2f;
+	private double deactivationTime = 2f;
 	private boolean disableDeactivation = false;
-	
+
 	public static ContactAddedCallback getContactAddedCallback() {
 		return threadLocal.get().gContactAddedCallback;
 	}
@@ -88,22 +88,22 @@ public class BulletGlobals {
 	public static void setContactProcessedCallback(ContactProcessedCallback callback) {
 		threadLocal.get().gContactProcessedCallback = callback;
 	}
-	
+
 	////////////////////////////////////////////////////////////////////////////
 
-	public static float getContactBreakingThreshold() {
+	public static double getContactBreakingThreshold() {
 		return threadLocal.get().contactBreakingThreshold;
 	}
 
-	public static void setContactBreakingThreshold(float threshold) {
+	public static void setContactBreakingThreshold(double threshold) {
 		threadLocal.get().contactBreakingThreshold = threshold;
 	}
 
-	public static float getDeactivationTime() {
+	public static double getDeactivationTime() {
 		return threadLocal.get().deactivationTime;
 	}
 
-	public static void setDeactivationTime(float time) {
+	public static void setDeactivationTime(double time) {
 		threadLocal.get().deactivationTime = time;
 	}
