+++ src/main/java/com/bulletphysics/collision/broadphase/AxisSweep3.java
@@ -1,6 +1,6 @@
 /*
  * Java port of Bullet (c) 2008 Martin Dvorak <jezek2@advel.cz>
- * 
+ *
  * AxisSweep3
  * Copyright (c) 2006 Simon Hobbs
  *
@@ -10,11 +10,11 @@
  * This software is provided 'as-is', without any express or implied warranty.
  * In no event will the authors be held liable for any damages arising from
  * the use of this software.
- * 
- * Permission is granted to anyone to use this software for any purpose, 
+ *
+ * Permission is granted to anyone to use this software for any purpose,
  * including commercial applications, and to alter it and redistribute it
  * freely, subject to the following restrictions:
- * 
+ *
  * 1. The origin of this software must not be misrepresented; you must not
  *    claim that you wrote the original software. If you use this software
  *    in a product, an acknowledgment in the product documentation would be
@@ -26,13 +26,13 @@
 
 package com.bulletphysics.collision.broadphase;
 
-import javax.vecmath.Vector3f;
+import javax.vecmath.Vector3d;
 
 /**
  * AxisSweep3 is an efficient implementation of the 3D axis sweep and prune broadphase.<p>
- * 
+ *
  * It uses arrays rather then lists for storage of the 3 axis. Also it operates using 16 bit
- * integer coordinates instead of floats. For large worlds and many objects, use {@link AxisSweep3_32}
+ * integer coordinates instead of doubles. For large worlds and many objects, use {@link AxisSweep3_32}
  * instead. AxisSweep3_32 has higher precision and allows more than 16384 objects at the cost
  * of more memory and bit of performance.
  *
@@ -40,20 +40,20 @@ import javax.vecmath.Vector3f;
  */
 public class AxisSweep3 extends AxisSweep3Internal {
 
-	public AxisSweep3(Vector3f worldAabbMin, Vector3f worldAabbMax) {
+	public AxisSweep3(Vector3d worldAabbMin, Vector3d worldAabbMax) {
 		this(worldAabbMin, worldAabbMax, 16384, null);
 	}
 
-	public AxisSweep3(Vector3f worldAabbMin, Vector3f worldAabbMax, int maxHandles) {
+	public AxisSweep3(Vector3d worldAabbMin, Vector3d worldAabbMax, int maxHandles) {
 		this(worldAabbMin, worldAabbMax, maxHandles, null);
 	}
-	
-	public AxisSweep3(Vector3f worldAabbMin, Vector3f worldAabbMax, int maxHandles/* = 16384*/, OverlappingPairCache pairCache/* = 0*/) {
+
+	public AxisSweep3(Vector3d worldAabbMin, Vector3d worldAabbMax, int maxHandles/* = 16384*/, OverlappingPairCache pairCache/* = 0*/) {
 		super(worldAabbMin, worldAabbMax, 0xfffe, 0xffff, maxHandles, pairCache);
 		// 1 handle is reserved as sentinel
 		assert (maxHandles > 1 && maxHandles < 32767);
 	}
-	
+
 	@Override
 	protected EdgeArray createEdgeArray(int size) {
 		return new EdgeArrayImpl(size);
@@ -63,11 +63,11 @@ public class AxisSweep3 extends AxisSweep3Internal {
 	protected Handle createHandle() {
 		return new HandleImpl();
 	}
-	
+
 	protected int getMask() {
 		return 0xFFFF;
 	}
-	
+
 	protected static class EdgeArrayImpl extends EdgeArray {
 		private short[] pos;
 		private short[] handle;
@@ -76,25 +76,25 @@ public class AxisSweep3 extends AxisSweep3Internal {
 			pos = new short[size];
 			handle = new short[size];
 		}
-		
+
 		@Override
 		public void swap(int idx1, int idx2) {
 			short tmpPos = pos[idx1];
 			short tmpHandle = handle[idx1];
-			
+
 			pos[idx1] = pos[idx2];
 			handle[idx1] = handle[idx2];
-			
+
 			pos[idx2] = tmpPos;
 			handle[idx2] = tmpHandle;
 		}
-		
+
 		@Override
 		public void set(int dest, int src) {
 			pos[dest] = pos[src];
 			handle[dest] = handle[src];
 		}
-		
+
 		@Override
 		public int getPos(int index) {
 			return pos[index] & 0xFFFF;
@@ -115,7 +115,7 @@ public class AxisSweep3 extends AxisSweep3Internal {
 			handle[index] = (short)value;
 		}
 	}
-	
+
 	protected static class HandleImpl extends Handle {
 		private short minEdges0;
 		private short minEdges1;
@@ -124,7 +124,7 @@ public class AxisSweep3 extends AxisSweep3Internal {
 		private short maxEdges0;
 		private short maxEdges1;
 		private short maxEdges2;
-		
+
 		@Override
 		public int getMinEdges(int edgeIndex) {
 			switch (edgeIndex) {
@@ -134,7 +134,7 @@ public class AxisSweep3 extends AxisSweep3Internal {
 				case 2: return minEdges2 & 0xFFFF;
 			}
 		}
-		
+
 		@Override
 		public void setMinEdges(int edgeIndex, int value) {
 			switch (edgeIndex) {
@@ -143,7 +143,7 @@ public class AxisSweep3 extends AxisSweep3Internal {
 				case 2: minEdges2 = (short)value; break;
 			}
 		}
-		
+
 		@Override
 		public int getMaxEdges(int edgeIndex) {
 			switch (edgeIndex) {
@@ -153,7 +153,7 @@ public class AxisSweep3 extends AxisSweep3Internal {
 				case 2: return maxEdges2 & 0xFFFF;
 			}
 		}
-		
+
 		@Override
 		public void setMaxEdges(int edgeIndex, int value) {
 			switch (edgeIndex) {
