+++ src/main/java/com/bulletphysics/collision/broadphase/AxisSweep3_32.java
@@ -1,6 +1,6 @@
 /*
  * Java port of Bullet (c) 2008 Martin Dvorak <jezek2@advel.cz>
- * 
+ *
  * AxisSweep3
  * Copyright (c) 2006 Simon Hobbs
  *
@@ -10,11 +10,11 @@
  * This software is provided 'as-is', without any express or implied warranty.
  * In no event will the authors be held liable for any damages arising from
  * the use of this software.
- * 
- * Permission is granted to anyone to use this software for any purpose, 
+ *
+ * Permission is granted to anyone to use this software for any purpose,
  * including commercial applications, and to alter it and redistribute it
  * freely, subject to the following restrictions:
- * 
+ *
  * 1. The origin of this software must not be misrepresented; you must not
  *    claim that you wrote the original software. If you use this software
  *    in a product, an acknowledgment in the product documentation would be
@@ -26,7 +26,7 @@
 
 package com.bulletphysics.collision.broadphase;
 
-import javax.vecmath.Vector3f;
+import javax.vecmath.Vector3d;
 
 /**
  * AxisSweep3_32 allows higher precision quantization and more objects compared
@@ -37,20 +37,20 @@ import javax.vecmath.Vector3f;
  */
 public class AxisSweep3_32 extends AxisSweep3Internal {
 
-	public AxisSweep3_32(Vector3f worldAabbMin, Vector3f worldAabbMax) {
+	public AxisSweep3_32(Vector3d worldAabbMin, Vector3d worldAabbMax) {
 		this(worldAabbMin, worldAabbMax, 1500000, null);
 	}
 
-	public AxisSweep3_32(Vector3f worldAabbMin, Vector3f worldAabbMax, int maxHandles) {
+	public AxisSweep3_32(Vector3d worldAabbMin, Vector3d worldAabbMax, int maxHandles) {
 		this(worldAabbMin, worldAabbMax, maxHandles, null);
 	}
-	
-	public AxisSweep3_32(Vector3f worldAabbMin, Vector3f worldAabbMax, int maxHandles/* = 1500000*/, OverlappingPairCache pairCache/* = 0*/) {
+
+	public AxisSweep3_32(Vector3d worldAabbMin, Vector3d worldAabbMax, int maxHandles/* = 1500000*/, OverlappingPairCache pairCache/* = 0*/) {
 		super(worldAabbMin, worldAabbMax, 0xfffffffe, 0x7fffffff, maxHandles, pairCache);
 		// 1 handle is reserved as sentinel
 		assert (maxHandles > 1 && maxHandles < 2147483647);
 	}
-	
+
 	@Override
 	protected EdgeArray createEdgeArray(int size) {
 		return new EdgeArrayImpl(size);
@@ -60,11 +60,11 @@ public class AxisSweep3_32 extends AxisSweep3Internal {
 	protected Handle createHandle() {
 		return new HandleImpl();
 	}
-	
+
 	protected int getMask() {
 		return 0xFFFFFFFF;
 	}
-	
+
 	protected static class EdgeArrayImpl extends EdgeArray {
 		private int[] pos;
 		private int[] handle;
@@ -73,25 +73,25 @@ public class AxisSweep3_32 extends AxisSweep3Internal {
 			pos = new int[size];
 			handle = new int[size];
 		}
-		
+
 		@Override
 		public void swap(int idx1, int idx2) {
 			int tmpPos = pos[idx1];
 			int tmpHandle = handle[idx1];
-			
+
 			pos[idx1] = pos[idx2];
 			handle[idx1] = handle[idx2];
-			
+
 			pos[idx2] = tmpPos;
 			handle[idx2] = tmpHandle;
 		}
-		
+
 		@Override
 		public void set(int dest, int src) {
 			pos[dest] = pos[src];
 			handle[dest] = handle[src];
 		}
-		
+
 		@Override
 		public int getPos(int index) {
 			return pos[index];
@@ -112,7 +112,7 @@ public class AxisSweep3_32 extends AxisSweep3Internal {
 			handle[index] = value;
 		}
 	}
-	
+
 	protected static class HandleImpl extends Handle {
 		private int minEdges0;
 		private int minEdges1;
@@ -121,7 +121,7 @@ public class AxisSweep3_32 extends AxisSweep3Internal {
 		private int maxEdges0;
 		private int maxEdges1;
 		private int maxEdges2;
-		
+
 		@Override
 		public int getMinEdges(int edgeIndex) {
 			switch (edgeIndex) {
@@ -131,7 +131,7 @@ public class AxisSweep3_32 extends AxisSweep3Internal {
 				case 2: return minEdges2;
 			}
 		}
-		
+
 		@Override
 		public void setMinEdges(int edgeIndex, int value) {
 			switch (edgeIndex) {
@@ -140,7 +140,7 @@ public class AxisSweep3_32 extends AxisSweep3Internal {
 				case 2: minEdges2 = value; break;
 			}
 		}
-		
+
 		@Override
 		public int getMaxEdges(int edgeIndex) {
 			switch (edgeIndex) {
@@ -150,7 +150,7 @@ public class AxisSweep3_32 extends AxisSweep3Internal {
 				case 2: return maxEdges2;
 			}
 		}
-		
+
 		@Override
 		public void setMaxEdges(int edgeIndex, int value) {
 			switch (edgeIndex) {
