+++ src/main/java/com/bulletphysics/collision/broadphase/CollisionAlgorithm.java
@@ -7,11 +7,11 @@
  * This software is provided 'as-is', without any express or implied warranty.
  * In no event will the authors be held liable for any damages arising from
  * the use of this software.
- * 
- * Permission is granted to anyone to use this software for any purpose, 
+ *
+ * Permission is granted to anyone to use this software for any purpose,
  * including commercial applications, and to alter it and redistribute it
  * freely, subject to the following restrictions:
- * 
+ *
  * 1. The origin of this software must not be misrepresented; you must not
  *    claim that you wrote the original software. If you use this software
  *    in a product, an acknowledgment in the product documentation would be
@@ -32,16 +32,16 @@ import com.bulletphysics.util.ObjectArrayList;
 /**
  * Collision algorithm for handling narrowphase or midphase collision detection
  * between two collision object types.
- * 
+ *
  * @author jezek2
  */
 public abstract class CollisionAlgorithm {
 
 	//protected final BulletStack stack = BulletStack.get();
-	
+
 	// JAVA NOTE: added
 	private CollisionAlgorithmCreateFunc createFunc;
-	
+
 	protected Dispatcher dispatcher;
 
 	public void init() {
@@ -50,15 +50,15 @@ public abstract class CollisionAlgorithm {
 	public void init(CollisionAlgorithmConstructionInfo ci) {
 		dispatcher = ci.dispatcher1;
 	}
-	
+
 	public abstract void destroy();
 
 	public abstract void processCollision(CollisionObject body0, CollisionObject body1, DispatcherInfo dispatchInfo, ManifoldResult resultOut);
 
-	public abstract float calculateTimeOfImpact(CollisionObject body0, CollisionObject body1, DispatcherInfo dispatchInfo, ManifoldResult resultOut);
-	
+	public abstract double calculateTimeOfImpact(CollisionObject body0, CollisionObject body1, DispatcherInfo dispatchInfo, ManifoldResult resultOut);
+
 	public abstract void getAllContactManifolds(ObjectArrayList<PersistentManifold> manifoldArray);
-	
+
 	public final void internalSetCreateFunc(CollisionAlgorithmCreateFunc func) {
 		createFunc = func;
 	}
@@ -66,5 +66,5 @@ public abstract class CollisionAlgorithm {
 	public final CollisionAlgorithmCreateFunc internalGetCreateFunc() {
 		return createFunc;
 	}
-	
+
 }
